Moved configurableoption into it's own file
